Split statesync modules
