+++ monad-statesync/src/progress.rs
@@ -1,0 +1,89 @@
+// Copyright (C) 2025 Category Labs, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+use alloy_consensus::Header;
+use monad_executor_glue::StateSyncRequest;
+use monad_types::SeqNum;
+
+const NUM_PREFIXES: u64 = 256;
+
+#[derive(Clone, Copy, Default)]
+pub(crate) struct StateSyncProgress {
+    start_target: Option<SeqNum>,
+    end_target: Option<SeqNum>,
+
+    // our guess for the minimum block at which servicers statesync'd before
+    min_until_guess: Option<SeqNum>,
+    current_progress: Option<u64>,
+}
+
+impl StateSyncProgress {
+    pub fn update_target(&mut self, target: &Header) {
+        self.end_target = Some(SeqNum(target.number));
+        if self.min_until_guess.is_none() {
+            // guess that the statesync servicers have been up for at least 10_000 blocks
+            self.min_until_guess = Some(SeqNum(target.number.max(10_000) - 10_000));
+        }
+    }
+
+    pub fn update_handled_request(&mut self, request: &StateSyncRequest) {
+        assert_eq!(self.end_target, Some(SeqNum(request.target)));
+        let min_until_guess = self.min_until_guess.expect("self.end_target exists").0;
+
+        if self.start_target.is_none() {
+            self.start_target = Some(SeqNum(request.from));
+        }
+        let start_target = self.start_target.expect("start_target set").0;
+
+        if self.current_progress.is_none() {
+            self.current_progress = Some(request.from * NUM_PREFIXES);
+        }
+
+        if request.until >= min_until_guess {
+            let adjusted_from = if request.from <= min_until_guess {
+                start_target
+            } else {
+                request.from
+            };
+            *self
+                .current_progress
+                .as_mut()
+                .expect("current_progress was set") += request.until - adjusted_from;
+        }
+    }
+
+    pub fn update_reached_target(&mut self, target: &Header) {
+        assert_eq!(self.end_target, Some(SeqNum(target.number)));
+        self.start_target = Some(SeqNum(target.number));
+        self.current_progress = None;
+    }
+
+    pub fn estimate(&self) -> Option<SeqNum> {
+        let start_target = self.start_target?;
+        let end_target = self.end_target?;
+
+        if start_target == end_target {
+            return Some(end_target);
+        }
+
+        assert!(end_target > start_target);
+
+        let _total_progress = (end_target - start_target).0 * NUM_PREFIXES;
+        // current_progress / _total_progress would estimate progress in percentage terms
+
+        // current_progress / num_prefixes can be used as a target estimate
+        Some(SeqNum(self.current_progress? / NUM_PREFIXES))
+    }
+}
